+++ CHANGES.md
@@ -76,6 +76,7 @@ New:
   renamed `normalize.old`.
 - New and better `compress` function. The previous one was renamed
   `compress.old` (#868, #869).
+- Added `stereo.width`.
 
 Changed: